requirements.txt
# Telegram Bot Dependencies
aiogram==3.13.1
telethon==1.36.0
aiofiles==24.1.0
python-dotenv==1.0.1
aiosqlite==0.20.0
cryptography==42.0.8
prettytable==3.10.0

# Web Admin Panel Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing & Export
pandas>=2.2.0
openpyxl>=3.1.2
reportlab>=4.0.7
xlsxwriter>=3.1.9

# HTTP Client for webhooks
httpx>=0.25.2

# Validation & Serialization
pydantic>=2.5.0
email-validator>=2.1.0

# Background Tasks & Scheduling
apscheduler>=3.10.4

# Monitoring & Logging
structlog>=23.2.0

# Development & Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
